requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asgiproxy==0.2.0
attrs==25.3.0
bidict==0.23.1
boto3==1.40.40
botocore==1.40.40
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
databricks-sdk==0.67.0
databricks-sqlalchemy==2.0.1
distro==1.9.0
fastapi==0.118.0
frozenlist==1.7.0
git-remote-s3==0.2.5
google-auth==2.40.3
granian==2.5.4
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jmespath==1.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.6.4
narwhals==2.6.0
numpy==2.3.3
openai
packaging==25.0
pandas==2.3.2
pgvector==0.4.1
platformdirs==4.4.0
plotly==6.3.0
propcache==0.3.2
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
reflex==0.8.13a1
reflex-ag-grid==0.0.11
reflex-enterprise==0.3.4.post2
reflex-global-hotkey==1.2.3
reflex-hosting-cli==0.1.56
requests==2.32.5
rich==14.1.0
rsa==4.9.1
s3transfer==0.14.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1